requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
requests>=2.31.0
perplexityai>=0.0.1
fastmcp>=2.0.0
httpx>=0.28.0

